requirements.txt
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
mcp>=1.0.0
boto3>=1.28.0
google-cloud>=0.34.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
aiofiles>=23.0.0
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0
click>=8.0.0
pyjwt>=2.0.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
